+++ examples/flights/src/AppArrow.svelte
@@ -1,18 +1,17 @@
 <script lang="ts">
 	import View1DHist from "./components/View1DHist.svelte";
 	import View2DHeat from "./components/View2DHeat.svelte";
-	import { Falcon } from "../../../falcon2/src/core/falcon";
+	import { Falcon, ArrowDB } from "../../../falcon2/src";
+	import { tableFromIPC } from "apache-arrow";
+	import { onMount } from "svelte";
+	import logo from "../../../logo/logo.png";
 	import type {
 		View0DState,
 		View1DState,
 		View2DState,
 		View1D,
 		View2D,
-	} from "../../../falcon2/src/core/views";
-	import { ArrowDB } from "../../../falcon2/src/db/arrow";
-	import { tableFromIPC } from "apache-arrow";
-	import { onMount } from "svelte";
-	import logo from "../../../logo/logo.png";
+	} from "../../../falcon2/src";
 
 	let totalCountState: View0DState;
 	let distanceState: View1DState;
