+++ examples/flights/src/main.ts
@@ -2,7 +2,8 @@
  * Write a js export wrapper here for the components at some point
  */
 // import App from "./App.svelte";
-import App from "./AppRefactor.svelte";
+// import App from "./AppArrow.svelte";
+import App from "./AppDuckDB.svelte";
 
 const app = new App({
 	target: document.getElementById("app"),
