+++ falcon2/src/index.ts
@@ -1,6 +1,3 @@
-export * from "./api";
-export * from "./basic";
-export * from "./config";
-export * from "./db/index";
-export * from "./util";
-export * from "./core/index";
+export * from "./core/views";
+export * from "./core/falcon";
+export * from "./old/db";
